Move OSVersionHelper from Common.UI up into ManagedCommon
