+++ src/common/ManagedCommon/OSVersionHelper.cs
@@ -4,10 +4,15 @@
 
 using System;
 
-namespace Common.UI
+namespace ManagedCommon
 {
     public static class OSVersionHelper
     {
+        public static bool IsWindows10()
+        {
+            return Environment.OSVersion.Version.Major >= 10 && Environment.OSVersion.Version.Minor < 22000;
+        }
+
         public static bool IsWindows11()
         {
             return Environment.OSVersion.Version.Major >= 10 && Environment.OSVersion.Version.Build >= 22000;
