+++ src/common/ManagedCommon/WindowHelpers.cs
@@ -40,12 +40,15 @@ namespace ManagedCommon
         /// Workaround for a WinUI bug on Windows 10 in which a window's top border is always
         /// black. Calls <c>DwmExtendFrameIntoClientArea()</c> with a <c>cyTopHeight</c> of 2 to force
         /// the window's top border to be visible.<br/><br/>
-        /// Has no visible effect on Windows 11.
+        /// Is a no-op on versions other than Windows 10.
         /// </summary>
-        public static void ForceTopBorder1PixelInset(IntPtr handle)
+        public static void ForceTopBorder1PixelInsetOnWindows10(IntPtr handle)
         {
-            var margins = new NativeMethods.MARGINS { cxLeftWidth = 0, cxRightWidth = 0, cyBottomHeight = 0, cyTopHeight = 2 };
-            NativeMethods.DwmExtendFrameIntoClientArea(handle, ref margins);
+            if (OSVersionHelper.IsWindows10())
+            {
+                var margins = new NativeMethods.MARGINS { cxLeftWidth = 0, cxRightWidth = 0, cyBottomHeight = 0, cyTopHeight = 2 };
+                NativeMethods.DwmExtendFrameIntoClientArea(handle, ref margins);
+            }
         }
     }
 }
