+++ src/modules/AdvancedPaste/AdvancedPaste/AdvancedPasteXAML/MainWindow.xaml.cs
@@ -82,7 +82,7 @@ namespace AdvancedPaste
             };
 
             WindowHelpers.BringToForeground(this.GetWindowHandle());
-            WindowHelpers.ForceTopBorder1PixelInset(this.GetWindowHandle());
+            WindowHelpers.ForceTopBorder1PixelInsetOnWindows10(this.GetWindowHandle());
         }
 
         private void OnActivated(object sender, WindowActivatedEventArgs args)
