+++ src/modules/Hosts/Hosts/HostsXAML/MainWindow.xaml.cs
@@ -39,7 +39,7 @@ namespace Hosts
 
             var handle = this.GetWindowHandle();
 
-            WindowHelpers.ForceTopBorder1PixelInset(handle);
+            WindowHelpers.ForceTopBorder1PixelInsetOnWindows10(handle);
             WindowHelpers.BringToForeground(handle);
             Activated += MainWindow_Activated;
