+++ src/settings-ui/Settings.UI/ViewModels/FancyZonesViewModel.cs
@@ -4,9 +4,8 @@
 
 using System;
 using System.Runtime.CompilerServices;
-
-using Common.UI;
 using global::PowerToys.GPOWrapper;
+using ManagedCommon;
 using Microsoft.PowerToys.Settings.UI.Library;
 using Microsoft.PowerToys.Settings.UI.Library.Helpers;
 using Microsoft.PowerToys.Settings.UI.Library.Interfaces;
