+++ src/settings-ui/Settings.UI/ViewModels/PowerOcrViewModel.cs
@@ -9,9 +9,8 @@ using System.Globalization;
 using System.Linq;
 using System.Text.Json;
 using System.Timers;
-
-using Common.UI;
 using global::PowerToys.GPOWrapper;
+using ManagedCommon;
 using Microsoft.PowerToys.Settings.UI.Library;
 using Microsoft.PowerToys.Settings.UI.Library.Helpers;
 using Microsoft.PowerToys.Settings.UI.Library.Interfaces;
